requirements.txt
streamlit==1.32.2
pypdf==4.1.0
langchain==0.1.13
python-dotenv==1.0.1
chromadb==0.4.24
langchain-openai==0.1.7